Clean up StudentInfo page & components.
