+++ client/src/pages/student-info/components/CriteriaCharts.tsx
@@ -5,23 +5,16 @@ import { ScheinCriteriaSummary } from 'shared/model/ScheinCriteria';
 import ChartPaper from '../../../components/info-paper/ChartPaper';
 
 interface Props extends GridProps {
-  firstCard?: React.ReactNode;
   scheinStatus: ScheinCriteriaSummary;
 }
 
-function CriteriaCharts({ scheinStatus, firstCard, ...props }: Props): JSX.Element {
+function CriteriaCharts({ scheinStatus, ...props }: Props): JSX.Element {
   const theme = useTheme();
 
   return (
-    <Grid container spacing={2} {...props}>
-      {firstCard && (
-        <Grid item sm={12} md={6} lg={4}>
-          {firstCard}
-        </Grid>
-      )}
-
+    <>
       {Object.values(scheinStatus.scheinCriteriaSummary).map((summary) => (
-        <Grid item key={summary.id} sm={12} md={6} lg={4}>
+        <Grid {...props} item key={summary.id}>
           <ChartPaper
             title={summary.name}
             chartType='PieChart'
@@ -43,7 +36,7 @@ function CriteriaCharts({ scheinStatus, firstCard, ...props }: Props): JSX.Eleme
           />
         </Grid>
       ))}
-    </Grid>
+    </>
   );
 }
